// convex/_generated/server.d.ts
/* eslint-disable */
/**
 * Generated server utilities.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * Generated by convex@1.26.1.
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ArgsArray,
  DefaultFunctionArgs,
  FunctionArgs,
  FunctionReference,
  FunctionReturnType,
  OptionalRestArgs,
  RegisteredMutation,
  RegisteredQuery,
  RegisteredAction,
} from "convex/server";

export declare const mutation: <
  ArgsValidator extends DefaultFunctionArgs,
  Output
>(
  func: {
    args?: ArgsValidator;
    handler: (ctx: any, args: FunctionArgs<ArgsValidator>) => Promise<Output>;
  }
) => FunctionReference<"mutation", "public", ArgsValidator, Output>;

export declare const query: <
  ArgsValidator extends DefaultFunctionArgs,
  Output
>(
  func: {
    args?: ArgsValidator;
    handler: (ctx: any, args: FunctionArgs<ArgsValidator>) => Promise<Output>;
  }
) => FunctionReference<"query", "public", ArgsValidator, Output>;

export declare const action: <
  ArgsValidator extends DefaultFunctionArgs,
  Output
>(
  func: {
    args?: ArgsValidator;
    handler: (ctx: any, args: FunctionArgs<ArgsValidator>) => Promise<Output>;
  }
) => FunctionReference<"action", "public", ArgsValidator, Output>;